Generalize the estimated heigth process so it will be able to work with other source than OSM
