+++ processingchain/src/main/groovy/org/orbisgis/orbisprocess/geoclimate/processingchain/GeoIndicatorsChain.groovy
@@ -1,6 +1,8 @@
 package org.orbisgis.orbisprocess.geoclimate.processingchain
 
 import groovy.transform.BaseScript
+import org.apache.commons.io.FileUtils
+import org.apache.commons.io.FilenameUtils
 import org.orbisgis.orbisdata.datamanager.jdbc.JdbcDataSource
 import org.orbisgis.orbisdata.processmanager.api.IProcess
 import org.orbisgis.orbisdata.processmanager.process.GroovyProcessFactory
@@ -994,8 +996,7 @@ IProcess createUnitsOfAnalysis() {
 /**
  * Compute all geoindicators at the 3 scales :
  * building, block and RSU
- * Compute also the LCZ classification (using the min distance algorithm - lczRandomForest=false -
- * or the randomForest one - lczRandomForest=true) and the urban typology
+ * Compute also the LCZ classification and the urban typology
  *
  * @return 4 tables outputTableBuildingIndicators, outputTableBlockIndicators, outputTableRsuIndicators,
  * outputTableRsuLcz . The first three tables contains the geoindicators and the last table the LCZ classification.
@@ -1013,8 +1014,8 @@ IProcess computeAllGeoIndicators() {
                 mapOfWeights: ["sky_view_factor"             : 1, "aspect_ratio": 1, "building_surface_fraction": 1,
                                "impervious_surface_fraction" : 1, "pervious_surface_fraction": 1,
                                "height_of_roughness_elements": 1, "terrain_roughness_length": 1],
-                lczRandomForest: false, lczModelName: "LCZ_OSM_RF_1_0.model",
-                urbanTypoModelName: "URBAN_TYPOLOGY_BDTOPO_V2_RF_2_0.model"
+                urbanTypoModelName: "",
+                buildingHeightModelName: ""
         outputs outputTableBuildingIndicators: String, outputTableBlockIndicators: String,
                 outputTableRsuIndicators: String, outputTableRsuLcz: String, outputTableZone: String,
                 outputTableRsuUrbanTypoArea: String, outputTableRsuUrbanTypoFloorArea: String,
@@ -1022,27 +1023,454 @@ IProcess computeAllGeoIndicators() {
         run { datasource, zoneTable, buildingTable, roadTable, railTable, vegetationTable, hydrographicTable,
               imperviousTable,
               surface_vegetation, surface_hydro, distance, indicatorUse, svfSimplified, prefixName, mapOfWeights,
-              lczRandomForest, lczModelName, urbanTypoModelName ->
+              urbanTypoModelName, buildingHeightModelName ->
+            //Estimate height
+            if (buildingHeightModelName) {
+                info "Geoclimate will try to estimate the building heights with the model $buildingHeightModelName."
+                //Let's check if the model exists
+                File inputModelFile = new File(buildingHeightModelName)
+                def modelName = FilenameUtils.getBaseName(buildingHeightModelName)
+                if (!inputModelFile.exists()) {
+                    //We try to find this model in geoclimate
+                    def modelURL = "https://github.com/orbisgis/geoclimate/raw/master/models/${modelName}.model"
+                    def localInputModelFile = new File(System.getProperty("user.home") + File.separator + ".geoclimate" + File.separator + modelName + ".model")
+                    // The model doesn't exist on the local folder we download it
+                    if (!localInputModelFile.exists()) {
+                        FileUtils.copyURLToFile(new URL(modelURL), localInputModelFile)
+                        if (!localInputModelFile.exists()) {
+                            error "Cannot find any model file to estimate the building heights"
+                            return
+                        }
+                    }
+                } else {
+                    if(!FilenameUtils.isExtension(pathAndFileName, "model")){
+                        error "The extension of the model file must be .model"
+                        return
+                    }
+                }
+                IProcess geoIndicators =  computeGeoclimateIndicators()
+                if (!geoIndicators.execute(datasource: datasource, zoneTable: zoneTable,
+                        buildingTable: buildingTable,
+                        roadTable: roadTable,
+                        railTable: railTable, vegetationTable: vegetationTable,
+                        hydrographicTable: hydrographicTable, imperviousTable: imperviousTable,
+                        surface_vegetation: surface_vegetation, surface_hydro: surface_hydro,
+                        indicatorUse:["URBAN_TYPOLOGY"],
+                        svfSimplified: true, prefixName: prefixName,
+                        mapOfWeights: mapOfWeights,
+                        urbanTypoModelName:"")){
+                    error "Cannot build the geoindicators to estimate the building height"
+                    return
+                }
+                //Let's go to select the building's id that must be processed to fix the height
+                info "Extracting the building having no height information and estimate it"
+                def results = geoIndicators.results;
+                //Select indicators we need at building scales
+                def buildingIndicatorsTableName = results.outputTableBuildingIndicators;
+                datasource.getTable(buildingEstimateTableName).id_build.createIndex()
+                datasource.getTable(buildingIndicatorsTableName).id_build.createIndex()
+                datasource.getTable(buildingIndicatorsTableName).id_rsu.createIndex()
+
+                def estimated_building_with_indicators = "ESTIMATED_BUILDING_INDICATORS_${UUID.randomUUID().toString().replaceAll("-", "_")}"
+
+                datasource.execute """DROP TABLE IF EXISTS $estimated_building_with_indicators;
+                                           CREATE TABLE $estimated_building_with_indicators 
+                                                    AS SELECT a.*
+                                                    FROM $buildingIndicatorsTableName a 
+                                                        RIGHT JOIN $buildingEstimateTableName b 
+                                                        ON a.id_build=b.id_build
+                                                    WHERE b.ESTIMATED = true AND a.ID_RSU IS NOT NULL;"""
+
+                info "Collect building indicators to estimate the height"
+
+                def applygatherScales = Geoindicators.GenericIndicators.gatherScales()
+                applygatherScales.execute([
+                        buildingTable    : estimated_building_with_indicators,
+                        blockTable       : results.outputTableBlockIndicators,
+                        rsuTable         : results.outputTableRsuIndicators,
+                        targetedScale    : "BUILDING",
+                        operationsToApply: ["AVG", "STD"],
+                        prefixName       : processing_parameters.prefixName,
+                        datasource       : datasource])
+                def gatheredScales = applygatherScales.results.outputTableName
+
+                info "Start estimating the building height"
+
+                //Apply RF model
+                def applyRF = Geoindicators.TypologyClassification.applyRandomForestModel()
+                applyRF.execute([
+                        explicativeVariablesTableName: gatheredScales,
+                        pathAndFileName              : buildingHeightModelName,
+                        idName                       : "id_build",
+                        prefixName                   : processing_parameters.prefixName,
+                        datasource                   : datasource])
+
+                //Update the abstract building table
+                info "Replace the input building table by the estimated height"
+                def buildEstimatedHeight = applyRF.results.outputTableName
+
+                datasource.getTable(buildEstimatedHeight).id_build.createIndex()
+
+                def newEstimatedHeigthWithIndicators = "NEW_BUILDING_INDICATORS_${UUID.randomUUID().toString().replaceAll("-", "_")}"
+
+                //Use build table indicators
+                datasource.execute """DROP TABLE IF EXISTS $newEstimatedHeigthWithIndicators;
+                                           CREATE TABLE $newEstimatedHeigthWithIndicators as 
+                                            SELECT  a.THE_GEOM, a.ID_BUILD,a.ID_SOURCE,
+                                        CASE WHEN b.HEIGHT_ROOF IS NULL THEN a.HEIGHT_WALL ELSE 0 END AS HEIGHT_WALL ,
+                                                COALESCE(b.HEIGHT_ROOF, a.HEIGHT_ROOF) AS HEIGHT_ROOF,
+                                                CASE WHEN b.HEIGHT_ROOF IS NULL THEN a.NB_LEV ELSE 0 END AS NB_LEV, a.TYPE,a.MAIN_USE, a.ZINDEX, a.ID_BLOCK, a.ID_RSU from $buildingIndicatorsTableName
+                                        a LEFT JOIN $buildEstimatedHeight b on a.id_build=b.id_build"""
+
+                //We must format only estimated buildings
+                //Apply format on the new abstract table
+                IProcess formatEstimatedBuilding = OSM.formatEstimatedBuilding
+                formatEstimatedBuilding.execute([
+                        datasource                : datasource,
+                        inputTableName            : newEstimatedHeigthWithIndicators,
+                        epsg                      : srid])
+                def newbuildingTableName = formatEstimatedBuilding.results.outputTableName
+
+                //Drop tables
+                datasource.execute """DROP TABLE IF EXISTS $estimated_building_with_indicators,
+                                        $newEstimatedHeigthWithIndicators, $buildEstimatedHeight,
+                                        $gatheredScales, $buildingTableName"""
+
+                //We use the existing spatial units
+                def relationBlocks = results.outputTableBlockIndicators
+                def relationRSU  = results.outputTableRsuIndicators
+
+                //The spatial relation tables RSU and BLOCK  must be filtered to keep only necessary columns
+                def rsuRelationFiltered = prefix prefixName, "RSU_RELATION_"
+                datasource.execute """DROP TABLE IF EXISTS $rsuRelationFiltered;
+            CREATE TABLE $rsuRelationFiltered AS SELECT ID_RSU, THE_GEOM FROM $relationRSU;
+            DROP TABLE $relationRSU;"""
+
+                def relationBlocksFiltered = prefix prefixName, "BLOCK_RELATION_"
+                datasource.execute """DROP TABLE IF EXISTS $relationBlocksFiltered;
+            CREATE TABLE $relationBlocksFiltered AS SELECT ID_BLOCK,  THE_GEOM,ID_RSU FROM $relationBlocks;
+            DROP TABLE $relationBlocks;"""
+
+                // Temporary (and output tables) are created
+                def lczIndicTable = postfix "LCZ_INDIC_TABLE"
+                def baseNameUrbanTypoRsu = prefix prefixName, "URBAN_TYPO_RSU_"
+                def urbanTypoBuilding
+                def distribNotPercent = "DISTRIB_NOT_PERCENT"
+                def COLUMN_ID_RSU = "id_rsu"
+                def COLUMN_ID_BUILD = "id_build"
+                def GEOMETRIC_COLUMN = "the_geom"
+                def CORRESPONDENCE_TAB_URB_TYPO = ["ba": 1,"bgh": 2,"icif": 3,"icio": 4,"id": 5,"local": 6,"pcif": 7,
+                                                   "pcio": 8,"pd": 9,"psc": 10]
+                def nameColTypoMaj = "TYPO_MAJ"
+
+                // Output Lcz (and urbanTypo) table names are set to null in case LCZ indicators (and urban typo) are not calculated
+                def rsuLcz = null
+                def urbanTypoArea = baseNameUrbanTypoRsu + "AREA"
+                def urbanTypoFloorArea = baseNameUrbanTypoRsu + "FLOOR_AREA"
+                def rsuLczWithoutGeom = "rsu_lcz_without_geom"
+
+                //Compute building indicators
+                def computeBuildingsIndicators = ProcessingChain.GeoIndicatorsChain.computeBuildingsIndicators()
+                if (!computeBuildingsIndicators.execute([datasource            : datasource,
+                                                         inputBuildingTableName: relationBuildings,
+                                                         inputRoadTableName    : roadTable,
+                                                         indicatorUse          : indicatorUse,
+                                                         prefixName            : prefixName])) {
+                    error "Cannot compute the building indicators"
+                    return null
+                }
+
+                def buildingIndicators = computeBuildingsIndicators.results.outputTableName
+
+                //Compute block indicators
+                def blockIndicators = null
+                if (indicatorUse*.toUpperCase().contains("URBAN_TYPOLOGY")) {
+                    def computeBlockIndicators = ProcessingChain.GeoIndicatorsChain.computeBlockIndicators()
+                    if (!computeBlockIndicators.execute([datasource            : datasource,
+                                                         inputBuildingTableName: buildingIndicators,
+                                                         inputBlockTableName   : relationBlocksFiltered,
+                                                         prefixName            : prefixName])) {
+                        error "Cannot compute the block indicators"
+                        return null
+                    }
+                    blockIndicators = computeBlockIndicators.results.outputTableName
+                }
+
+                //Compute RSU indicators
+                def rsuIndicators = null
+                def computeRSUIndicators = ProcessingChain.GeoIndicatorsChain.computeRSUIndicators()
+                if (!computeRSUIndicators.execute([datasource       : datasource,
+                                                   buildingTable    : buildingIndicators,
+                                                   rsuTable         : rsuRelationFiltered,
+                                                   vegetationTable  : vegetationTable,
+                                                   roadTable        : roadTable,
+                                                   hydrographicTable: hydrographicTable,
+                                                   imperviousTable  : imperviousTable,
+                                                   indicatorUse     : indicatorUse,
+                                                   svfSimplified    : svfSimplified,
+                                                   prefixName       : prefixName])) {
+                    error "Cannot compute the RSU indicators"
+                    return null
+                }
+                rsuIndicators = computeRSUIndicators.results.outputTableName
+                info "All geoindicators have been computed"
+
+                // If the LCZ indicators should be calculated, we only affect a LCZ class to each RSU
+                if (indicatorUse.contains("LCZ")) {
+                    info """ The LCZ classification is performed """
+
+                    def lczIndicNames = ["GEOM_AVG_HEIGHT_ROOF"              : "HEIGHT_OF_ROUGHNESS_ELEMENTS",
+                                             "BUILDING_FRACTION_LCZ"             : "BUILDING_SURFACE_FRACTION",
+                                             "ASPECT_RATIO"                      : "ASPECT_RATIO",
+                                             "GROUND_SKY_VIEW_FACTOR"            : "SKY_VIEW_FACTOR",
+                                             "PERVIOUS_FRACTION_LCZ"             : "PERVIOUS_SURFACE_FRACTION",
+                                             "IMPERVIOUS_FRACTION_LCZ"           : "IMPERVIOUS_SURFACE_FRACTION",
+                                             "EFFECTIVE_TERRAIN_ROUGHNESS_LENGTH": "TERRAIN_ROUGHNESS_LENGTH"]
+
+                        // Get into a new table the ID, geometry column and the 7 indicators defined by Stewart and Oke (2012)
+                        // for LCZ classification (rename the indicators with the real names)
+                        def queryReplaceNames = ""
+                        lczIndicNames.each { oldIndic, newIndic ->
+                            queryReplaceNames += "ALTER TABLE $lczIndicTable ALTER COLUMN $oldIndic RENAME TO $newIndic;"
+                        }
+                        datasource.execute """DROP TABLE IF EXISTS $lczIndicTable;
+                                CREATE TABLE $lczIndicTable 
+                                        AS SELECT $COLUMN_ID_RSU, $GEOMETRIC_COLUMN, ${lczIndicNames.keySet().join(",")} 
+                                        FROM ${computeRSUIndicators.results.outputTableName};
+                                $queryReplaceNames"""
+
+                        datasource."$lczIndicTable".reload()
+
+                        // The classification algorithm is called
+                        def classifyLCZ = Geoindicators.TypologyClassification.identifyLczType()
+                        if (!classifyLCZ([rsuLczIndicators : lczIndicTable,
+                                          rsuAllIndicators : computeRSUIndicators.results.outputTableName,
+                                          normalisationType: "AVG",
+                                          mapOfWeights     : mapOfWeights,
+                                          prefixName       : prefixName,
+                                          datasource       : datasource,
+                                          prefixName       : prefixName])) {
+                            info "Cannot compute the LCZ classification."
+                            return
+                        }
+                        rsuLcz = classifyLCZ.results.outputTableName
+                        datasource.execute "DROP TABLE IF EXISTS $lczIndicTable"
+
+                }
+                // If the URBAN_TYPOLOGY indicators should be calculated, we only affect a URBAN typo class
+                // to each building and then to each RSU
+                if (indicatorUse.contains("URBAN_TYPOLOGY") && urbanTypoModelName) {
+                    info """ The URBAN TYPOLOGY classification is performed """
+                    applygatherScales = Geoindicators.GenericIndicators.gatherScales()
+                    applygatherScales.execute([
+                            buildingTable    : buildingIndicators,
+                            blockTable       : blockIndicators,
+                            rsuTable         : rsuIndicators,
+                            targetedScale    : "BUILDING",
+                            operationsToApply: ["AVG", "STD"],
+                            prefixName       : prefixName,
+                            datasource       : datasource])
+                    gatheredScales = applygatherScales.results.outputTableName
+
+                    applyRF = Geoindicators.TypologyClassification.applyRandomForestModel()
+                    applyRF.execute([
+                            explicativeVariablesTableName: gatheredScales,
+                            pathAndFileName              : urbanTypoModelName,
+                            idName                       : COLUMN_ID_BUILD,
+                            prefixName                   : prefixName,
+                            datasource                   : datasource])
+                    def urbanTypoBuild = applyRF.results.outputTableName
+
+                    // Creation of a list which contains all types of the urban typology (in their string version)
+                    def urbTypoCorrespondenceTabInverted = [:]
+                    CORRESPONDENCE_TAB_URB_TYPO.each{fin, ini->
+                        urbTypoCorrespondenceTabInverted[ini]=fin
+                    }
+                    datasource."$urbanTypoBuild".I_TYPO.createIndex()
+                    def queryDistinct = """SELECT DISTINCT I_TYPO AS I_TYPO FROM $urbanTypoBuild"""
+                    def mapTypos = datasource.rows(queryDistinct)
+                    def listTypos = []
+                    mapTypos.each{
+                        listTypos.add(urbTypoCorrespondenceTabInverted[it.I_TYPO])
+                    }
+
+                    // Join the geometry field to the building typology table and replace integer by string values
+                    def queryCaseWhenReplace = ""
+                    def endCaseWhen = ""
+                    urbTypoCorrespondenceTabInverted.each{ini, fin ->
+                        queryCaseWhenReplace += "CASE WHEN b.I_TYPO=$ini THEN '$fin' ELSE "
+                        endCaseWhen += " END"
+                    }
+                    queryCaseWhenReplace = queryCaseWhenReplace + " 'unknown' " + endCaseWhen
+                    urbanTypoBuilding = prefix  prefixName, "URBAN_TYPO_BUILDING"
+                    datasource."$urbanTypoBuild"."$COLUMN_ID_BUILD".createIndex()
+                    datasource."$buildingIndicators"."$COLUMN_ID_BUILD".createIndex()
+                    datasource """  DROP TABLE IF EXISTS $urbanTypoBuilding;
+                                CREATE TABLE $urbanTypoBuilding
+                                    AS SELECT   a.$COLUMN_ID_BUILD, a.$COLUMN_ID_RSU, a.THE_GEOM,
+                                                $queryCaseWhenReplace AS I_TYPO
+                                    FROM $buildingIndicators a LEFT JOIN $urbanTypoBuild b
+                                    ON a.$COLUMN_ID_BUILD = b.$COLUMN_ID_BUILD
+                                    WHERE a.$COLUMN_ID_RSU IS NOT NULL"""
+
+                    // Create a distribution table (for each RSU, contains the % area OR floor area of each urban typo)
+                    def queryCasewhen = [:]
+                    queryCasewhen["AREA"]=""
+                    queryCasewhen["FLOOR_AREA"]=""
+                    queryCasewhen.keySet().each{ind ->
+                        def querySum = ""
+                        listTypos.each{typoCol ->
+                            queryCasewhen[ind] += """ SUM(CASE WHEN a.I_TYPO='$typoCol' THEN b.$ind ELSE 0 END) AS TYPO_$typoCol,"""
+                            querySum = querySum + " COALESCE(b.TYPO_${typoCol}/(b.TYPO_${listTypos.join("+b.TYPO_")}), 0) AS TYPO_$typoCol,"
+                        }
+                        // Calculates the distribution per RSU
+                        datasource."$buildingIndicators"."$COLUMN_ID_RSU".createIndex()
+                        datasource."$urbanTypoBuilding"."$COLUMN_ID_BUILD".createIndex()
+                        datasource.execute """  DROP TABLE IF EXISTS $distribNotPercent;
+                                            CREATE TABLE $distribNotPercent
+                                                AS SELECT   b.$COLUMN_ID_RSU,
+                                                            ${queryCasewhen[ind][0..-2]} 
+                                                FROM $urbanTypoBuilding a RIGHT JOIN $buildingIndicators b
+                                                ON a.$COLUMN_ID_BUILD = b.$COLUMN_ID_BUILD
+                                                WHERE b.$COLUMN_ID_RSU IS NOT NULL 
+                                                GROUP BY b.$COLUMN_ID_RSU
+                                                """
+                        // Calculates the frequency by RSU
+                        datasource."$distribNotPercent"."$COLUMN_ID_RSU".createIndex()
+                        datasource."$rsuIndicators"."$COLUMN_ID_RSU".createIndex()
+                        datasource.execute """  DROP TABLE IF EXISTS TEMPO_DISTRIB;
+                                            CREATE TABLE TEMPO_DISTRIB
+                                                AS SELECT   a.$COLUMN_ID_RSU, a.the_geom,
+                                                            ${querySum[0..-2]} 
+                                                FROM $rsuIndicators a LEFT JOIN $distribNotPercent b
+                                                ON a.$COLUMN_ID_RSU = b.$COLUMN_ID_RSU"""
+
+                        // Characterize the distribution to identify the most frequent type within a RSU
+                        def computeDistribChar = Geoindicators.GenericIndicators.distributionCharacterization()
+                        computeDistribChar([distribTableName: "TEMPO_DISTRIB",
+                                            inputId         : COLUMN_ID_RSU,
+                                            initialTable    : "TEMPO_DISTRIB",
+                                            distribIndicator: ["uniqueness"],
+                                            extremum        : "GREATEST",
+                                            keep2ndCol      : false,
+                                            keepColVal      : false,
+                                            prefixName      : "${prefixName}$ind",
+                                            datasource      : datasource])
+                        def resultsDistrib = computeDistribChar.results.outputTableName
+
+                        // Join main typo table with distribution table and replace typo by null when it has been set
+                        // while there is no building in the RSU
+                        datasource."$resultsDistrib"."$COLUMN_ID_RSU".createIndex()
+                        datasource.tempo_distrib."$COLUMN_ID_RSU".createIndex()
+                        datasource """  DROP TABLE IF EXISTS $baseNameUrbanTypoRsu$ind;
+                                    CREATE TABLE $baseNameUrbanTypoRsu$ind
+                                        AS SELECT   a.*, 
+                                                    CASE WHEN   b.UNIQUENESS_VALUE=-1
+                                                    THEN        NULL
+                                                    ELSE        b.UNIQUENESS_VALUE END AS UNIQUENESS_VALUE,
+                                                    CASE WHEN   b.UNIQUENESS_VALUE=-1
+                                                    THEN        NULL
+                                                    ELSE        LOWER(SUBSTRING(b.EXTREMUM_COL FROM 6)) END AS $nameColTypoMaj
+                                        FROM    TEMPO_DISTRIB a LEFT JOIN $resultsDistrib b
+                                        ON a.$COLUMN_ID_RSU=b.$COLUMN_ID_RSU"""
+                    }
+                    // Drop temporary tables
+                    datasource """DROP TABLE IF EXISTS $urbanTypoBuild, $gatheredScales, $distribNotPercent, TEMPO_DISTRIB"""
+                }
+                else{
+                    urbanTypoArea = null
+                    urbanTypoFloorArea = null
+                    urbanTypoBuilding = null
+                }
+
+                datasource.execute "DROP TABLE IF EXISTS $rsuLczWithoutGeom;"
+
+                return [outputTableBuildingIndicators   : computeBuildingsIndicators.getResults().outputTableName,
+                        outputTableBlockIndicators      : blockIndicators,
+                        outputTableRsuIndicators        : computeRSUIndicators.getResults().outputTableName,
+                        outputTableRsuLcz               : rsuLcz,
+                        outputTableZone                 : zoneTable,
+                        outputTableRsuUrbanTypoArea     : urbanTypoArea,
+                        outputTableRsuUrbanTypoFloorArea: urbanTypoFloorArea,
+                        outputTableBuildingUrbanTypo    : urbanTypoBuilding]
+
+            }
+            else{
+                IProcess geoIndicators =  computeGeoclimateIndicators()
+                if (!geoIndicators.execute(datasource: datasource, zoneTable: zoneTable,
+                        buildingTable: buildingTable,
+                        roadTable: roadTable,
+                        railTable: railTable, vegetationTable: vegetationTable,
+                        hydrographicTable: hydrographicTable, imperviousTable: imperviousTable,
+                        surface_vegetation: surface_vegetation, surface_hydro: surface_hydro,
+                        indicatorUse:indicatorUse,
+                        svfSimplified: svfSimplified, prefixName: prefixName,
+                        mapOfWeights: mapOfWeights,
+                        urbanTypoModelName:urbanTypoModelName)){
+                    error "Cannot build the geoindicators"
+                    return
+                }
+                return geoIndicators.getResults()
+                /*return [outputTableBuildingIndicators   :results.outputTableBuildingIndicators,
+                        outputTableBlockIndicators      : results,
+                        outputTableRsuIndicators        : results,
+                        outputTableRsuLcz               : results,
+                        outputTableZone                 : results,
+                        outputTableRsuUrbanTypoArea     : results,
+                        outputTableRsuUrbanTypoFloorArea: results,
+                        outputTableBuildingUrbanTypo    : results]*/
+            }
+        }
+    }
+}
+
+/**
+ * Compute all geoclimate indicators at the 3 scales :
+ * building, block and RSU
+ * The LCZ classification  and the urban typology
+ *
+ * @return 8 tables outputTableBuildingIndicators, outputTableBlockIndicators, outputTableRsuIndicators,
+ * outputTableRsuLcz, outputTableZone ,
+ * outputTableRsuUrbanTypoArea, outputTableRsuUrbanTypoFloorArea,
+ * outputTableBuildingUrbanTypo.
+ * The first three tables contains the geoindicators and the last tables the LCZ and urban typology classifications.
+ * This table can be empty if the user decides not to calculate it.
+ *
+ */
+IProcess computeGeoclimateIndicators() {
+    return create {
+        title "Compute all geoindicators"
+        id "computeAllGeoIndicators"
+        inputs datasource: JdbcDataSource, zoneTable: "", buildingTable: "",
+                roadTable: "", railTable: "", vegetationTable: "",
+                hydrographicTable: "", imperviousTable: "", surface_vegetation: 10000, surface_hydro: 2500,
+                distance: 0.01, indicatorUse: ["LCZ", "URBAN_TYPOLOGY", "TEB"], svfSimplified: false, prefixName: "",
+                mapOfWeights: ["sky_view_factor"             : 1, "aspect_ratio": 1, "building_surface_fraction": 1,
+                               "impervious_surface_fraction" : 1, "pervious_surface_fraction": 1,
+                               "height_of_roughness_elements": 1, "terrain_roughness_length": 1],
+                urbanTypoModelName: ""
+        outputs outputTableBuildingIndicators: String, outputTableBlockIndicators: String,
+                outputTableRsuIndicators: String, outputTableRsuLcz: String, outputTableZone: String,
+                outputTableRsuUrbanTypoArea: String, outputTableRsuUrbanTypoFloorArea: String,
+                outputTableBuildingUrbanTypo: String
+        run { datasource, zoneTable, buildingTable, roadTable, railTable, vegetationTable, hydrographicTable,
+              imperviousTable,
+              surface_vegetation, surface_hydro, distance, indicatorUse, svfSimplified, prefixName, mapOfWeights,
+               urbanTypoModelName ->
             info "Start computing the geoindicators..."
+
             // Temporary (and output tables) are created
             def lczIndicTable = postfix "LCZ_INDIC_TABLE"
             def baseNameUrbanTypoRsu = prefix prefixName, "URBAN_TYPO_RSU_"
             def urbanTypoBuilding
             def distribNotPercent = "DISTRIB_NOT_PERCENT"
 
-
             def COLUMN_ID_RSU = "id_rsu"
             def COLUMN_ID_BUILD = "id_build"
             def GEOMETRIC_COLUMN = "the_geom"
-            def CORRESPONDENCE_TAB_URB_TYPO = ["ba": 1,"bgh": 2,"icif": 3,"icio": 4,"id": 5,"local": 6,"pcif": 7,
-                                               "pcio": 8,"pd": 9,"psc": 10]
+            def CORRESPONDENCE_TAB_URB_TYPO = ["ba"  : 1, "bgh": 2, "icif": 3, "icio": 4, "id": 5, "local": 6, "pcif": 7,
+                                               "pcio": 8, "pd": 9, "psc": 10]
             def nameColTypoMaj = "TYPO_MAJ"
-
-            // If the randomForest should be used, need to calculate all indicators
-            if (indicatorUse*.toUpperCase().contains("LCZ") && lczRandomForest) {
-                indicatorUse = ["URBAN_TYPOLOGY", "LCZ"]
-            }
-
             //Check data before computing indicators
             if (!zoneTable && !buildingTable && !roadTable) {
                 error "To compute the geoindicators the zone, building and road tables must not be null or empty"
@@ -1087,8 +1515,7 @@ IProcess computeAllGeoIndicators() {
 
             //Compute block indicators
             def blockIndicators = null
-            if ((indicatorUse*.toUpperCase().contains("URBAN_TYPOLOGY")) ||
-                    (indicatorUse*.toUpperCase().contains("LCZ") && lczRandomForest)) {
+            if (indicatorUse*.toUpperCase().contains("URBAN_TYPOLOGY")) {
                 def computeBlockIndicators = computeBlockIndicators()
                 if (!computeBlockIndicators.execute([datasource            : datasource,
                                                      inputBuildingTableName: buildingIndicators,
@@ -1122,38 +1549,7 @@ IProcess computeAllGeoIndicators() {
             // If the LCZ indicators should be calculated, we only affect a LCZ class to each RSU
             if (indicatorUse.contains("LCZ")) {
                 info """ The LCZ classification is performed """
-                if (lczRandomForest) {
-                    def applygatherScales = Geoindicators.GenericIndicators.gatherScales()
-                    applygatherScales.execute([
-                            buildingTable    : buildingIndicators,
-                            blockTable       : blockIndicators,
-                            rsuTable         : rsuIndicators,
-                            targetedScale    : "RSU",
-                            operationsToApply: ["AVG", "STD"],
-                            prefixName       : prefixName,
-                            datasource       : datasource])
-                    def gatheredScales = applygatherScales.results.outputTableName
-
-                    def applyRF = Geoindicators.TypologyClassification.applyRandomForestModel()
-                    applyRF.execute([
-                            explicativeVariablesTableName: gatheredScales,
-                            pathAndFileName              : lczModelName,
-                            idName                       : COLUMN_ID_RSU,
-                            prefixName                   : prefixName,
-                            datasource                   : datasource])
-                    rsuLczWithoutGeom = applyRF.results.outputTableName
-                    datasource.execute """  ALTER TABLE $rsuLczWithoutGeom RENAME COLUMN LCZ TO LCZ1;"""
-                    datasource."$rsuLczWithoutGeom".reload()
-                    datasource."$rsuLczWithoutGeom"."$COLUMN_ID_RSU".createIndex()
-                    datasource."$relationRSU"."$COLUMN_ID_RSU".createIndex()
-                    rsuLcz = prefix(prefixName, "rsu_lcz")
-                    datasource.execute """  DROP TABLE IF EXISTS $rsuLcz;
-                                            CREATE TABLE $rsuLcz
-                                                    AS SELECT a.*, b.the_geom
-                                                    FROM $rsuLczWithoutGeom a RIGHT JOIN $relationRSU b
-                                                    ON a.$COLUMN_ID_RSU = b.$COLUMN_ID_RSU"""
-                } else {
-                    def lczIndicNames = ["GEOM_AVG_HEIGHT_ROOF"              : "HEIGHT_OF_ROUGHNESS_ELEMENTS",
+                def lczIndicNames = ["GEOM_AVG_HEIGHT_ROOF"              : "HEIGHT_OF_ROUGHNESS_ELEMENTS",
                                          "BUILDING_FRACTION_LCZ"             : "BUILDING_SURFACE_FRACTION",
                                          "ASPECT_RATIO"                      : "ASPECT_RATIO",
                                          "GROUND_SKY_VIEW_FACTOR"            : "SKY_VIEW_FACTOR",
@@ -1189,7 +1585,7 @@ IProcess computeAllGeoIndicators() {
                     }
                     rsuLcz = classifyLCZ.results.outputTableName
                     datasource.execute "DROP TABLE IF EXISTS $lczIndicTable"
-                }
+
             }
             // If the URBAN_TYPOLOGY indicators should be calculated, we only affect a URBAN typo class
             // to each building and then to each RSU
@@ -1217,26 +1613,26 @@ IProcess computeAllGeoIndicators() {
 
                 // Creation of a list which contains all types of the urban typology (in their string version)
                 def urbTypoCorrespondenceTabInverted = [:]
-                CORRESPONDENCE_TAB_URB_TYPO.each{fin, ini->
-                    urbTypoCorrespondenceTabInverted[ini]=fin
+                CORRESPONDENCE_TAB_URB_TYPO.each { fin, ini ->
+                    urbTypoCorrespondenceTabInverted[ini] = fin
                 }
                 datasource."$urbanTypoBuild".I_TYPO.createIndex()
                 def queryDistinct = """SELECT DISTINCT I_TYPO AS I_TYPO FROM $urbanTypoBuild"""
                 def mapTypos = datasource.rows(queryDistinct)
                 def listTypos = []
-                mapTypos.each{
+                mapTypos.each {
                     listTypos.add(urbTypoCorrespondenceTabInverted[it.I_TYPO])
                 }
 
                 // Join the geometry field to the building typology table and replace integer by string values
                 def queryCaseWhenReplace = ""
                 def endCaseWhen = ""
-                urbTypoCorrespondenceTabInverted.each{ini, fin ->
+                urbTypoCorrespondenceTabInverted.each { ini, fin ->
                     queryCaseWhenReplace += "CASE WHEN b.I_TYPO=$ini THEN '$fin' ELSE "
                     endCaseWhen += " END"
                 }
                 queryCaseWhenReplace = queryCaseWhenReplace + " 'unknown' " + endCaseWhen
-                urbanTypoBuilding = prefix  prefixName, "URBAN_TYPO_BUILDING"
+                urbanTypoBuilding = prefix prefixName, "URBAN_TYPO_BUILDING"
                 datasource."$urbanTypoBuild"."$COLUMN_ID_BUILD".createIndex()
                 datasource."$buildingIndicators"."$COLUMN_ID_BUILD".createIndex()
                 datasource """  DROP TABLE IF EXISTS $urbanTypoBuilding;
@@ -1249,11 +1645,11 @@ IProcess computeAllGeoIndicators() {
 
                 // Create a distribution table (for each RSU, contains the % area OR floor area of each urban typo)
                 def queryCasewhen = [:]
-                queryCasewhen["AREA"]=""
-                queryCasewhen["FLOOR_AREA"]=""
-                queryCasewhen.keySet().each{ind ->
+                queryCasewhen["AREA"] = ""
+                queryCasewhen["FLOOR_AREA"] = ""
+                queryCasewhen.keySet().each { ind ->
                     def querySum = ""
-                    listTypos.each{typoCol ->
+                    listTypos.each { typoCol ->
                         queryCasewhen[ind] += """ SUM(CASE WHEN a.I_TYPO='$typoCol' THEN b.$ind ELSE 0 END) AS TYPO_$typoCol,"""
                         querySum = querySum + " COALESCE(b.TYPO_${typoCol}/(b.TYPO_${listTypos.join("+b.TYPO_")}), 0) AS TYPO_$typoCol,"
                     }
@@ -1311,8 +1707,7 @@ IProcess computeAllGeoIndicators() {
 
                 // Drop temporary tables
                 datasource """DROP TABLE IF EXISTS $urbanTypoBuild, $gatheredScales, $distribNotPercent, TEMPO_DISTRIB"""
-            }
-            else{
+            } else {
                 urbanTypoArea = null
                 urbanTypoFloorArea = null
                 urbanTypoBuilding = null
